Fix @ember-data/request-utils build
